- Restructure code into classes for less clutter.
