+++ xml_loader/rest_rule.py
@@ -1,0 +1,11 @@
+class Daily_rest_rule():
+    def __init__(self, hours, rest_id):
+        self.rest_id = rest_id
+        self.hours = hours
+        
+class Weekly_rest_rule():
+    def __init__(self, hours, rest_id):
+        self.rest_id = rest_id
+        self.hours = hours
+
+
